Only show spinner for runs that have not yet materialized
